% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Generates data from a linear model}
\usage{
dgp(nObs, nCov, nNonZero)
}
\arguments{
\item{nObs}{Number of observations}

\item{nCov}{Number of covariates, excluding intercept}

\item{nNonZero}{Number of covariates with coefficients set to one, from left
to right}
}
\description{
Generates data on where the outcome is the sum of nCov uniform(-1, 1) random
variables. All covariates are set to 0 by default, then a number corresponding
to nNonZero are changed to one.
}
