% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDesignMatrices.R
\name{generateDesignMatrices}
\alias{generateDesignMatrices}
\title{Generates design matrices from a dataset}
\usage{
generateDesignMatrices(complexity, sharedComplexity, dataSet)
}
\arguments{
\item{complexity}{Number of covariates in each model}

\item{sharedComplexity}{Number of covariates in common between both models}

\item{dataSet}{The data set used, the first column contains the response and
the is no included intercept}
}
\description{
Generates two design matrices, each with a specified complexity
(meaning number of covariates), and with a specific degree of 
shared complexity (meaning how many of their covariates
to they have in common).
}
\details{
The function builds the design matrices using up covariates from
the data set as needed starting from left to right. So for example,
if there is no shared complexity and each model has complexity 3 the
first design matrix will contain the three first covariates in the 
data set, and the second will contain covariates four through six. If 
there are more covariates than this in the data set these will be 
left unused. In the case where there is shared complexity, the covariates 
that are included in both models will be the first taken from left to 
right.

(Maybe a better way to do this would be to supply the function with 
two 0/1 vectors of which covariates to include.)
}
