% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocvLM.R
\name{loocvLM}
\alias{loocvLM}
\title{Calculates leave-one-out-cross-validation errors for a linear model}
\usage{
loocvLM(data, model)
}
\arguments{
\item{data}{The name of the data set}

\item{model}{The name of the (linear) model}
}
\description{
Calculates leave-one-out-cross-validation errors for a linear model
}
