% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Generates data from a linear model}
\usage{
dgp(n_obs, beta_coefs, true_mean = FALSE)
}
\arguments{
\item{n_obs}{Number of observations}

\item{beta_coefs}{Vector of beta coefficients. The length of the vector will
determine the number of covariates}

\item{trueMean}{If TRUE, returns response/outcome without noise. Defaults to
false}
}
\description{
Generates n_obs observations where the outcome (y) is the weighted sum, with
weights given by beta_coefs, of a number of independent uniform(-1, 1)
distributed random variables.
}
