% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Generates data from a linear model}
\usage{
dgp(nObs, nCov, nNonZero, trueMean = FALSE)
}
\arguments{
\item{nObs}{Number of observations}

\item{nCov}{Number of covariates, excluding intercept}

\item{nNonZero}{Number of covariates with coefficients set to one, from left
to right}

\item{trueMean}{If TRUE, returns response/outcome without noise. Defaults to
false}
}
\description{
Generates nObs observations where the outcome (y) is the sum of nNonZero  out of
nCov uniform(-1, 1) random variables. The covariates with non-zero coefficients
are always the nNonZero first covariates.
}
